pyproject.toml
[project]
name = "pgreviewer"
version = "0.1.0"
description = "Foundation of the pgreviewer project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "numpy>=2.4.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "rich>=13.0.0",
    "tqdm>=4.67.3",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-sql>=0.3.11",
    "typer>=0.9.0",
    "unidiff>=0.7.5",
]

[project.scripts]
pgr = "pgreviewer.cli.main:app"

[tool.setuptools.packages.find]
include = ["pgreviewer*"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=8.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=6.0",
    "ruff>=0.15.6",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "TCH", "PYI"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
source = ["pgreviewer"]
relative_files = true
omit = [
    "pgreviewer/cli/main.py",
    "pgreviewer/db/pool.py",
    "pgreviewer/analysis/explain_runner.py",
    "pgreviewer/infra/debug_store.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as requiring a live database connection (deselect with '-m \"not integration\"')",
]
